user have to complete a registration before using the bot
If an user for the course hasn't been configured, an email and password are asked.
if the credentials are right, then the user's credentials are stored in a table in a db (unencrypted).
All the users that choose a course with an account configured don't have to insert their credentials.
All the users' id, course and if they want to receive news are stored in a table, separated from the one with the credentials.
Newsletter not working.